requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
et_xmlfile==2.0.0
fastapi==0.121.3
h11==0.16.0
httptools==0.7.1
idna==3.11
joblib==1.5.2
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.6.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
